requirements.txt
locust
python-dotenv==1.1.0
